ShcoolLearner/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.3.7
python-jose==3.3.0
passlib==1.7.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.3.0
